requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg
django-cors-headers
django-jazzmin
python-dotenv
requests>=2.31
pillow
drf-spectacular
